Consistent whitespace in functions
The code was inconsistent, so I tried to make it consistent:
one space before the opening parenthesis in anonymous functions
and no space in named functions.
This is the convention in the project on which I am working
most of the time ( https://www.mediawiki.org/wiki/CC/JS ). If
Twitter has a different convention, feel free to change it.