Add support for more right-to-left scripts
The original version only supports Hebrew and Arabic. There are more
right-to-left scripts in current use, most importantly Thaana, N'Ko
and Syriac. I added support for all of them by making the character
range in the regular expression include all the relevant Unicode
blocks, which are consecutive in the UCS numbering.
Samaritan is probably used rarely, but it's in the same
consecutive Unicode range, so I included it, too.

I did not include other ancient scripts such as Phoenician
and Mandaic, because their use is rare and it's unlikely that they
would pose problems with punctuation, which are frequent on the web,
but for the sake of completeness it may be worth to include them, too.

For a (more or less) full list of all modern and ancient right-to-left
scripts supported by Unicode, see
https://en.wikipedia.org/wiki/Right-to-left